refacto: separate service tests into 2 files
